refactor: rating types

+++ QueueUp.Domain/Entities/EstablishmentRating.cs
@@ -3,7 +3,7 @@ namespace QueueUp.Domain.Entities;
 public class EstablishmentRating
 {
     public Guid Id { get; set; } = Guid.Empty;
-    public int Rating { get; set; } = 0;
+    public double Rating { get; set; } = 0;
     public DateTime CreatedAt { get; set; } = DateTime.Now;
     public Guid EstablishmentId { get; set; } = Guid.Empty;
     public Guid UserId { get; set; } = Guid.Empty;
